refactor: Break up scans into how they're executed to reduce confusion on how they're called.
